dayplanner-backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
alembic==1.15.1
psycopg2-binary==2.9.10
pydantic-settings==2.8.1
python-jose==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
httpx==0.28.1
pgvector==0.3.6
python-dotenv==1.0.1
pytest==8.3.5
pytest-asyncio==0.25.3
langgraph==0.2.67
langchain==0.3.12
langchain-openai==0.2.14
langchain-google-genai==2.0.7
